# Conflicts:
#	src/main/java/com/car_rental/car_rental/Services/LoanService.java
#	src/main/java/com/car_rental/car_rental/Services/RefundService.java
#	src/main/java/com/car_rental/car_rental/Services/ReservationService.java

+++ src/main/java/com/car_rental/car_rental/entites/Car.java
@@ -22,7 +22,7 @@ public class Car {
     private Long mileage;
     @Enumerated(EnumType.STRING)
     private Status status;
-    private Long amount;
+    private Double amount;
     @ManyToOne
     @JoinColumn(name = "branch_id")
     private Branch branch;
